Added ability to store phone numbers for students ...
For now, only Admins can add phone numbers. They will have to log-in as
the teacher, select students and feed in the phone numbers.

Some basic level checks/sanitization done on phone numbers ( using phony_rails ) gem.

Phone numbers are stored in the associated Account object - if available.
But when not ( only for shell Students ) - it is stored on the Student object
itself. On merge, phone number transferred to new Account