Made verticals/topics context aware
Whether or not to show # of questions and what the relevant count
is depends on the who is asking and in what context.

Only examiners are shown the # of questions on a topic.
And when the context is 'addhints', then only the # of questions authored
by them on that topic