A new model to track user's daily usage added ...
Called DAilyStreak, this model will be used to track the number of
points, marks or dollars earned for attempting problems - on a
day-to-day basis.

This is like the daily streak chart Duolingo generates.

First use is at the U&I event in DPS Bangalore where we will
use this table to track the total donations collected