Track weighted scores alongwith the nominal scores for a Skill ...
A question step or a snippet can have multiple skills attached to it.
Not all skills, however, carry equal weight.

In particular, skills from the same chapter to which the
question or snippet belongs carry twice the weight.

Hence, when a skill is attempted, its num_tested is incremented
by 1 - but its weighted_tested is increment by the skill's
weight in that snippet / question step.