[Big]: STI for Questions & Snippets + Skill-specific parceling
For historical reasons, we had separate models and tables
for Question and Snippets. Now, we have implemented STI
=> Question & Snippets run on top of a single Riddle table.

Also, made changes to parceling logic to allow questions
and snippets to be bundled into specific parcels for
specific skills. Will use these new parcels in the
prototype version 3.x of the app.